store: add blueprintEndpoints to composerApi
This commit moves the blueprint related endpoints to its own file and
extracts the helpers. We can also drop the migration script that copies
blueprints since RHEL10.2 has been released.
