+++ src/store/api/backend/onprem/composerApi/helpers/index.ts
@@ -1,1 +1,2 @@
+export { getBlueprintsPath } from './getBlueprintsPath';
 export { getCloudConfigs } from './getCloudConfigs';
